+++ app.vue
@@ -72,8 +72,8 @@ useHead({
       type: "image/webp",
       fetchpriority: "high",
     },
-    { rel: "preconnect", href: "https://cdn.userway.org" },
-    { rel: "dns-prefetch", href: "https://cdn.userway.org" },
+    { rel: "preconnect", href: "https://cdn.jsdelivr.net" },
+    { rel: "dns-prefetch", href: "https://cdn.jsdelivr.net" },
     { rel: "dns-prefetch", href: "https://www.googletagmanager.com" },
   ],
   script: [